fixes associated with renaming user model as seller
